refactor: ImageLoading component separation / loading skeleton when loading image
